Refactor to match traditional structure, update badges, README
