+++ .travis.yml
@@ -4,5 +4,6 @@ addons:
   apt:
     packages:
       - jblas
+      - libgfortran3
 after_success:
   - bash <(curl -s https://codecov.io/bash)
