+++ build.gradle
@@ -3,11 +3,11 @@ apply plugin: 'jacoco'
 apply plugin: 'application'
 apply plugin: 'com.github.johnrengelman.shadow'
 
-mainClassName = 'ca.craigthomas.visualclassifier.commandline.Runner'
-group = 'ca.craigthomas.visualclassifier'
+mainClassName = 'ca.craigthomas.neuralnetwork.commandline.Runner'
+group = 'ca.craigthomas.neuralnetwork'
 version = '1.0'
 
-description = 'A simple machine learning image classifier'
+description = 'A neural network for image classification'
 
 sourceCompatibility = 1.8
 
@@ -17,15 +17,15 @@ repositories {
 
 dependencies {
     compile 'com.beust:jcommander:1.72'
-    compile 'org.slf4j:slf4j-api:1.7.5'
+    compile 'org.slf4j:slf4j-api:1.7.28'
     compile 'org.boofcv:openkinect:0.17'
-    compile 'org.jblas:jblas:1.2.3'
+    compile 'org.jblas:jblas:1.2.4'
     compile 'org.apache.commons:commons-csv:1.0'
     compile 'commons-io:commons-io:2.4'
     compile 'org.apache.commons:commons-lang3:3.3.2'
-    compile 'org.apache.commons:commons-math3:3.3'
+    compile 'org.apache.commons:commons-math3:3.6.1'
     testCompile 'junit:junit:4.11'
-    testCompile 'org.mockito:mockito-all:1.9.5'
+    testCompile 'org.mockito:mockito-all:1.10.19'
 }
 
 buildscript {
@@ -35,23 +35,7 @@ buildscript {
     }
 
     dependencies {
-        classpath 'com.github.jengelman.gradle.plugins:shadow:2.0.2'
-    }
-}
-
-sourceSets {
-    main {
-        java {
-            srcDir 'src'
-        }
-    }
-    test {
-        java {
-            srcDir 'test'
-        }
-        resources {
-            srcDir 'resources'
-        }
+        classpath 'com.github.jengelman.gradle.plugins:shadow:5.1.0'
     }
 }
 
@@ -60,10 +44,6 @@ jacocoTestReport {
         xml.enabled = true
         html.enabled = true
     }
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = '4.5.1'
 }
 
 task stage {
