+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'visualclassifier'
+rootProject.name = 'neuralnetwork'
